AngularClassCompletion.cs and AngularLogoCompletion.cs content was swapped
